requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
google-generativeai
python-dotenv
gspread
oauth2client
pandas
